Added support for target-groups.
Target groups allow for steps to be part of multiple "virtual
targets" such as "all" or "full-teardown".

If a run targets a target group, terraflop will loop through the
step keys in each file repeatedly so as to run the targetted
step lists in that file in the same order as specified in the
target group.

There is currently no way using target-groups to run all of
target "a" in all files, then run target "b" in all files, etc.
For that, run `flop` multiple times, indicating an individual
target each time (e.g. flop a; flop b; ...)